Move "where" builder for locations to lib/utils.js, to allow reusing it.
