+++ lib/utils.js
@@ -1,8 +1,11 @@
 'use strict';
 
 import { forEach, pick, has, indexOf, isArray, omit, result, find, isFinite } from 'lodash';
+import config from 'config';
 import moment from 'moment';
 import countries from './country-list';
+
+const defaultGeoRadius = config.get('geoRadius');
 
 export function payloadToKey (namespace, payload) {
   //
@@ -224,6 +227,83 @@ export function buildSQLQuery (base, payload = {}, operators = [], betweens = []
   }
 
   return base;
+}
+
+/**
+ * Builds up a Knex where clause to query locations table.
+ * @params {object} query Query parameters passed to the endpoint.
+ * @return {object} Where clause to be used in a Knex query
+ */
+export function buildLocationsWhere (query) {
+  // Get used parameters
+  const {
+    country,
+    city,
+    location,
+    parameter,
+    has_geo,
+    coordinates,
+    radius
+  } = query;
+
+  return builder => {
+    if (country) {
+      builder.where('country', 'IN', [].concat(country));
+    }
+
+    if (city) {
+      // Ensure city is an array type, to perform map
+      const cities = [].concat(city);
+
+      // Transform "cities" field into string and search in it
+      builder.whereRaw(
+        cities
+          .map(c => "lower(array_to_string(cities,' ')) like ?")
+          .join(' OR '),
+        cities.map(c => `%${c.toLowerCase()}%`)
+      );
+    }
+
+    if (location) {
+      // Ensure location is an array, to perform map
+      const locations = [].concat(location);
+
+      // Transform "cities" field into string and search in it
+      builder.whereRaw(
+        locations
+          .map(l => "lower(array_to_string(locations,' ')) like ?")
+          .join(' OR '),
+        locations.map(l => `%${l.toLowerCase()}%`)
+      );
+    }
+
+    if (parameter) {
+      builder.where('parameters', '&&', [].concat(parameter));
+    }
+
+    /* eslint-disable camelcase */
+    if (typeof has_geo !== 'undefined') {
+      has_geo
+        ? builder.whereNotNull('coordinates')
+        : builder.whereNull('coordinates');
+    }
+    /* eslint-enable camelcase */
+
+    if (typeof radius !== 'undefined') {
+      if (coordinates) {
+        const [lat, lon] = coordinates.split(',');
+        builder.whereRaw(`ST_DWithin(
+            coordinates, 
+            ST_MakePoint(
+              ${parseFloat(lon)},${parseFloat(lat)}), 
+              ${radius || defaultGeoRadius}
+            )
+          `);
+      } else {
+        throw Error('"coordinates" must be passed with "radius" parameter.');
+      }
+    }
+  };
 }
 
 /**
